Get revisions even if the cluster was compacted.
Cluster may be compacted, thus the least available revision won't be 1. In order to receive the least available revision we do a dry run of watcher, it returns ErrCompacted (if requested compacted revision), which contains min revision. I hope for a better way to get this info, but there is no api in etcd to do so.

Also clear revisions table on continuous clicks on load button.